application: clamp CPU time to ensure monotonicity
redpanda_cpu_busy_seconds_total can go backwards (have a negative delta
from sample to sample) which is a problem because it is a counter
(despite the wrong type in the metadata) and this wreaks havoc on
prometheus due to violating the counter contract, causing false
“reset detect” and huge spikes in this metric.

This change clamps the value below by the highest seen value to date so
that the value never decreases. This only matters on totally unloaded
clusters and the negative values are very small and are generally
overcome the next period.

Fixes CORE-7909.