storage: Add new index_state test
The test checks that serialization output is binary compatible with the
current serialization. The test embeds the expected serialization output
to have something to compare with.

The motivation here is to be able to evolve segment index_state without
breaking the serialization/deserialization. The index_state will be
refactored in the following commits.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>