rptest: fail test if kgo verifier service is stopped before waited
This prevents incorrect assertions being performed on the status thread.
Stopping a service is a hard stop and does not guarantee that latest
status of the service was ingested. Instead, the user should call wait
first and optionally stop().

This prevents one particular problem where KGO did crash but the test
never noticed it because it did call stop() and wait() just does nothing
afterwards because it is expected that the service is down afterwards.