rptest: fix kgo stop(), wait() call order
This prevents incorrect assertions being performed on the status thread.
Stopping a service is a hard stop and does not guarantee that latest
status of the service was ingested. Instead, the user should call wait()
first and optionally stop().

If we call stop() first, then wait() this hides a problem in
DeleteRecordsTest where KGO crashes but the test never notices it
because stop() just killed the service and the status thread.

(cherry picked from commit 29f94d604dbbf6d78644a854b9ae89df873352ee)