rptest: remove invalid assert from DeleteRecordsTest
In immediately preceding commits I have fixed the order in which
consumer is shut down in this test which led which did hide a crash in
kgo verifier consumer. I have also fixed the crash in kgo.

This revealed a problem within DeleteRecordsTest and specifically the
assert on out_of_scope_invalid_reads. In this particular test, before
launching kgo producer, we first call produce_until_segments which uses
a kafka tools for producing data. These batches do not conform to what
kgo-verifier does expect. So it is perfectly valid to have
out_of_scope_invalid_reads.