rptest/htt: do not stop kgo before wait
wait() after stop() does not make sense because wait() is then short
circuited. In a previous commit I made this sequence of calls illegal to
prevent mistakes where the caller might have wrong beliefs about what
wait() does.

In this test it seems that it is never intended to wait for producer or
consumer to finish their work as we are producing an infinite (very
large number) amount of messages and the intent is just to generate load
so stop() is likely what was intended.

(cherry picked from commit c476f207ac935118ffa9a308f7b21d74ae5a691a)