dl/coordinator: background loop for committing files
Constructs an Iceberg file committer in the coordinator manager to be
used by the managed coordinators and runs a background loop that
periodically commits files.

The loop is structurally very similar to the NTP archival loop in that
it waits for leadership notifications and runs until the coordinator is
aborted. The loop calls into the now-instantiated Iceberg file committer
to build metadata updates, and then replicates the updates via the STM.