dt: fix fips CI hang on md5 usage
Using MD5 is not allowed in FIPS mode and so the default python
installation on a FIPS-enabled host prevents using MD5 without
explicitly marking it as `usedforsecurity=False`.

This has caused the nightly FIPS CI build to hang when the
`compute_storage.py` script is attempting to run md5.

This tries to fix the hanging test by marking the md5 usage with
`usedforsecurity=False`.