static_str: wraps a string with static lifetime
static_str wraps a constant string with known static ("indefinite")
lifetime. This useful because constructing and passing such strings from
character constants is very cheap: just a single "copy" of a
pointer-sized value, which can often even be elided. This opens the door
to using this in hot paths where an allocating string might raise
performance concerns.

In a subsequent series, this is used to enhance logging on the reader
cache path.

Relates to CORE-7865.