offline_log_viewer: allow reading newer versions of envelopes
Before this commit offline log viewer would refuse reading envelopes if
their version was higher than max_version. That is an overly restrictive
constraint. Our envelopes are designed to be forward and backwards
compatible so in most of the cases an older implementation of the log
viewer should be able to read newer versions of the envelopes just fine.

This commit implements that. Instead of putting a constraint on
version we constrain only compat_version like regular redpanda
deserialization would do.

Additional changes include issuing a warning if offline log viewer
compat version is lower than envelope version. This means that offline
log viewer will be able to decode only a subset of the envelope.

And obviously added code to skip over data that can't be read by the
offline log viewer.

Breaking changes in the envelopes are introduced by bumping compat
version. In that case offline log viewer will refuse the decoding as
expected.