batch_reader: expose flags for testing
It is currently difficult to test certain behaviors of the log reader
infrastructure, e.g., the interplay between the readers cache and the
batch readers.

In order to test some key aspects of this in our e2e storage test, we
introduce a get_flags() method on the private implementation, which
exposes (at the moment) details about the caching behavior of the
obtained reader. This doesn't expand the public interface of the reader,
only the private one which tests can get at.

CORE-7865.