storage: Refactor 'index_state'
This commit extracts three columns (relative offset, relative time, and
position indexes) out of the 'index_state' into its own component called
'index_columns'. The goal is to be able to swap the implementation of
the 'index_columns' with a different one.

The serialization format and general behavior didn't change. The main
difference is that instead of working with three collections directly
the code in the 'index_state' is now invoking methods of the
'index_columns'.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>