state_machine_base: redefine state_machine_base::apply()
This applies the batch and also sets the next offset to be applied under
a mutex. This ensures that by the time apply() returns, next offset is
updated too.

Competing actions that rely on the next offset being correct (like
snapshotting) also wait on the same lock so they see a consistent state
of the stm state.