dl/coordinator: construct Iceberg catalog
This constructs a default (filesystem) catalog in the coordinator
manager, so that it can be passed to the managed coordinators.

Longer term we will likely need a catalog factory of sorts to select
between different catalogs and configure them in more specific ways
(e.g. defining a base location for the cluster to write to).

For now though this commit just adds the plumbing necessary to get
a catalog through to the coordinator.