log_reader: return more readers to reader cache
In the log_reader, we sometimes incorrectly close the reader when it
should be marked for re-use. This causes historical reads which hit
their max_bytes condition (which is the common case for historical reads
as the only other ending condition is basically hitting the tip of the
log) to read 256K (128K demand read + 128K read-ahead) but drop the 128K
read-ahead portion of it, leading to 2x read amplification.

This occurs when the read from the last slice returns zero batches (a
common case when the batch size is large, e.g., of the same order of
magnitude of the 128K slice read size). In this case the load slice loop
closes the underlying segment iterator, due to the order of the checks,
even though the iterator is still usable, preventing caching.

This change fixes the order of the checks so that these readers will
returned to the cache.

Partial fix for CORE-7865.