log_reader: reset _last_base on reader cache hit
log_reader instances may be returned to the reader cache for use by
a subsequent reader: this helps preserve any read-ahead they have in
their ss::input_stream.

When a cache hit occurs the log_reader is "reset" to a state suitable
for the new reader to use it, but we did not reset the _last_base
field. This field is used in some progress checks during batch
reading, and it should be set to {} (default ctor value) for a fresh
or reset reader, or we may spuriously stop reading from the stream
and close the reader.

Along with prior commit this fixes CORE-7865.