dt: install license in fips environment
Running with fips enabled is an enterprise feature which requires a
license. So install a license in the cluster when running in a FIPS
environment to fix the currently failing FIPS CDT tests.

The license installing endpoint was introduced in 23.1, so we also skip
all tests that start from a version <23.1 in FIPS mode. These tests
could probably be deleted instead, but here we just skip them to limit
the scope of this PR. The exception is the `UpgradeBackToBackTest` which
does a rolling upgrade from a very old version. This seemed useful to
keep to test transitive backward compatibility, so here I just increased
the initial version for FIPS mode to be 23.1 to allow installing a
license.