storage: Implement columnar compression
Implement columnar compression for materialized indexes. The
compressed_index_columns class is a drop in replacement for the
index_columns which uses delta_for compression. This class is not used
directly but instaed the commit introduces index_columns_mux which is a
wrapper for the index_columns/compressed_index_columns. It uses
configuration parameter to decide what types of object to create and
redirects all calls to it.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>