dt: skip node uuid override tests in FIPS
These tests are currently failing because of how using an enterprise
feature (FIPS), using the UUID override functionality and the license
enforcement (upgrade blocking) interacts. Speficially, when we restart a
node with an overwritten UUID and a cleaned data directory:
1. we start with the logical version set to the initial version
2. we don't register with another node in the host (since that would
fail in leaderless cluster)
3. therefore we do not receive a controller snapshot to start with
4. which means that we won't have a license installed until later during
startup when the controller log is caught up
5. which means that license verification fails on startup because we
mistake this for an upgrade scenario (feature version is <
latest_version), fips is enabled, and there is not license installed.

Note that this implies that if we use the node UUID override feature in
an enterprise cluster, we likely also need to use the license
enforcement escape hatch (`REDPANDA_FALLBACK_ENTERPRISE_LICENSE`) as
well. This seems acceptable since they are both undocumented, expert
features, so here I skip the tests in FIPS mode instead of attempting to
address the underlying issue (eg. by waiting X seconds on startup before
the verification is performed when the UUID overrides are used.)