ct: Extract 'write_request' out of the batcher
This commit adds new nested directory ('core') which will contain all
code used by different subsystems in cloud topics (read path, write
path, resource management, etc).

The write_request and serializer are moved into the nested directory.
The write request will be generalized and used not only by the batcher
but also by the resource management subsystem and subsystem. The logic
in the batcher that manages in-flight write_request instances will be
generalized and used for other things and not only batcher.

This is needed to avoid code duplication between the susbsystems. For
instance, if the load balancing and throttling components are placed in
front of the batcher they will have to introduce their own
'write_request' analogs to manage in-flight requests.

Something similar has to be implemented for the read path but I don't
want to generalize write_request (and make it just request) to avoid
code bloat that will come with this. The goal of this change is to build
a scaffolding for the architecture without introducing too much new.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>