datalake: add schema_identifier
Adds a schema_identifier that uniquely identifies the structure of a
record component schema as it exists in the schema registry. For
instance, an Avro schema is identified by the PPSR schema ID; a protobuf
descriptor, on the other hand, is identified by the schema ID and a
vector of offsets. As such, the identifier contains and ID and an
optional vector.