datalake: add partitioning writer
Adds yet another abstraction to be used on the data path: the new
partitioning writer will be in charge of collecting records that
correspond to a single schema and grouping them by their appropriate
partition key.

The default schema will be to partition timestamps by the hour. A test
is added that demonstrates this.

This is not yet plugged into the record multiplexer, but the expectation
is that the multiplexer will construct one partitioning_writer per
Parquet schema.