rptest: assert consumer invariants after each upgrade
This improves failure result fidelity because we would spot the error
exactly after a bad upgrade rather than at the end of the test.

This also makes it simple(r) to skip validating old versions of
redpanda with newer versions of consumers which check additional
invariants.

This is to prepare for upgrading kgo to a version where it crashes
instantly on data loss on upgrade through some old versions of Redpanda
(around v21, v22). I plan to skip validating these versions and focus
only on latest supported releases instead.