storage/e2e_test: add reader caching checks
This adds a test case to the storage e2e test which checks that reader
caching will be effective under a variety of log reader scenarios,
in order to avoid something like CORE-7865 in the future. It
especially stresses reads around the 128K boundary.

We check that readers are re-usable as expected and also that reaeders
are obtained as cache hits as expected.

CORE-7865.