storage:log_reader: share reset logic with ctor
The log_reader offers a reset_config() method which is used by the
reader cache in order to re-use existing log_reader for new read
requests. This tries to set the log reader to a state similar to the
initial state that is set in the constructor, except that it keeps the
reader embedded in the iterator_pair around. The risk is that the reader
is not completely reset because reset doesn't reset "enough". This
happened with the _last_base member which was not reset inside
reset_config().

In order to reduce the chance of this happening, have the constructor
and reset_config() both call a private reset() method which does most of
the work.

CORE-7865.