storage:log_reader: move abort handling into reset
The log_reader can have an abort source passed in by the caller, which
can be used to abort the read in progress. The case reset_config() is
called, we do update the abort source in the reader so aborts will be
ineffective.

Move the logic from the constructor into reset() which is shared
between ctor and reset_config.