cluster: avoid duplicate partition registration callbacks
If registration occurs while a partition is co_await starting, it will
be present in the ntp_table and the upcalls will happen both inline with
registration and again after the partition is started.

This patch skips callback at time of registration if the partition is in
the process of starting, and relies on the callback that is invoked
after the partition is started.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>