ct: Use 'write_pipeline' in 'batcher'
Previously, the batcher was responsible for both housekeeping and
uploading. The housekeeping includes write_request creation and
accounting (including timeout handling).

The housekeeing is now moved to the write_pipeline. The goal is to be
able to extend the write path without adding new things to the batcher
itself. For instance, the throttler can now be impelemented
independently from the batcher. It could just consume the write_pipeline
as a dependency.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>