iceberg: schema method field id assignment
Adds a method to reassign the field IDs in a schema. The assignment
matches the output of the Java impl[1], which goes depth first into each
field, assigning IDs incrementally. Unlike the Java impl, this method
doesn't accept a set of existing fields IDs to preserve. This will
eventually be needed once we fully support schema evolution, but isn't
required for now.

[1] https://github.com/apache/iceberg/blob/main/api/src/main/java/org/apache/iceberg/types/AssignFreshIds.java