datalake: make record multiplexer write to multiple files
This commit fulfills the record_multiplexer's destiny in writing data
across multiple files. There are two dimension on which we will write
multiple files:
- by Iceberg partition key (the hour of the record timestamp)
- by the record value's schema, as resolved by the
  record_schema_resolver

The former is already encapsulated entirely within partitioning_writer,
so the record_multiplexer just needs to maintain a collection thereof.

A couple of companions have joined to help the record_multiplexer in its
journey: the schema manager, to register new schemas and fetch field
IDs; and the type resolver, to determine the schema from a given record.
These will likely be owned by the datalake manager, and for now, have
stub implementations where needed. A follow-up change will plumb the
schema registry and Iceberg catalog in so we can use the real
implementations.

A test is added with a dummy writer and Avro data to demonstrate the
multiplexing. A later commit will test with a Parquet writer.