datalake: add catalog schema manager
Adds a new component the catalog_schema_manager to get field IDs from
the Iceberg catalog, potentially registering a new schema if there are
missing fields in the catalog.

This will be the place we can make decisions about schema evolution. As
is, the implemented policy is that as long as fields are added to the
end of the table, the schema can be added. This is so that we can
support going from the default schema to the default schema + user
defined columns.

In the short term, for simplicity, I expect this to live along side the
datalake workers, where each worker will call into the schema manager
when constructing a parquet writer with a given schema to get the
correct field IDs. Longer term, for scale, perhaps this ought to live in
the coordinator and workers may interact with it via RPC.