iceberg: move manifest_io arg to transaction::merge_append
This allows transactions to be constructed without manifest_io, which
can be useful when callers know they are going to perform actions that
don't write manifests (e.g. updating the schema).

update_schema_action_test is updated as a good example of this, where
before we needlessly required s3 for this arg.