iceberg: assign fresh ids on schema update
Calls the new field id assignment method from the update schema action.
This makes the new schema contain field IDs that look like what is
generated by the Java impl[1].

When using the REST catalog, the catalogs ignore these and assign IDs on
its own with the Java impl. The filesystem catalog however will use the
field IDs as is. So this change effectively makes the two catalogs
result in the same field IDs.

[1] https://github.com/apache/iceberg/blob/main/api/src/main/java/org/apache/iceberg/types/AssignFreshIds.java