features: implement builtin evaluation period
Create a builtin evaluation period that allows customers to trial
Enterprise features for a trial period starting at cluster creation.

This is implemented as a real trial license created shortly after the
cluster is formed, when the cluster creation timestamp becomes
available. The controller leader coordinates the creation of the trial
license in the background.

The evaluation period license behaves like a real trial license was
installed shortly after cluster bootstrap. Therefore, it is visible to
the user through the admin API endpoints, license expiry metric,
prevents license nags, upgrade blocking and it is also backed up as a
real license for whole cluster recovery from cloud storage.