storage: remove assertion on `is_cloud_retention_active`
This assertion can be triggered if a the configuration for a topic is
changed between when space management sets the `_cloud_gc_offset` to
when gc actually runs on the log. So to avoid crashing in this case the
assertion is changed to an early return and a warning is logged instead.

(cherry picked from commit 137951eae3b4b9933848cdb1d0ff3911b29bc952)