serde/tests: fixed test validating parsing using incorrect schema
The test that used incorrect schema to parse a buffer serialized with
avro library, sometimes failed as the parsing didn't throw an error.
Changed the test to check if the behavior of `serde::parser` is the same
as the `::avro::GenericReader`. The test now validates that the outcome
of both parsers is the same.

Signed-off-by: Michał Maślanka <michal@redpanda.com>