c/partition_manager: re-download logs for half-initialized partitions
If a node goes down halfway through partition init we can end up with some
kvstore entries but without raft snapshot metadata. This leads to failed
assertions on the relative position of offsets.

The fix is to check for archival_metadata_stm snapshot presence before
downloading, rather than just for the directory.