iceberg: use `iceberg::uri` in iceberg metadata
Replaced type of all location fields that are expected to be
canonical uris with `iceberg::uri` type. In order to make this commit to
compile changed the `manifest_io` interface by introduction of overloads
accepting `uris`. The overloads will prevent us from making errors are
the right logic will be chosen based on the type provided.

Signed-off-by: Michał Maślanka <michal@redpanda.com>