datalake/tests: record generator
I want to write some tests that have some real-looking structured data,
with the magic byte + schema ID in values, and data serialized with Avro
(and in the future, protobuf).

To make tests a bit easier to write, and to hide PPSR schema ID as an
implementation detail, this generator keys schemas by a string that
tests will provide.