datalake: implement binary type resolver
Similar to the schema_manager abstract class, adds a type_resolver
abstract class and implements a simple binary type resolver that returns
a binary typed buffer.

Not only will this be useful to add a type_resolver to existing tests
without also adding a schema registry, but this will also be useful as
we will want to allow configuring an explicit key-value only schema
resolution mode.