datalake: add record_translator
Adds an abstraction that encapsulates translating data values into their
corresponding iceberg::value types. This is one of the places where
we'll define the user-facing schema for a given Protobuf/Avro schema.

As implemented, only a record's value is translated, and the translated
fields areappended to the end of the default schema. If there is no
translated type for the record value, the value buf is used to populate
the redpand_value field in the default schema.

No tests are added in this commit since this simply exercises existing
serialization code, but a more end-to-end test with the record
multiplexer will be added in a subsequent commit.