dt: ignore OIDC background errors in license tests
The default OIDC server is flaky in CI, so ignore the related error logs
in the background updater using `OIDC_UPDATE_FAILURE_LOGS`.

Alternatively we could create a local `KeyCloak` OIDC Authorization
Server which would be reliably available locally, but it's not simple to
configure, so it's simpler to ignore the error logs for the purpose of
these tests.