k/fetch: block audit topic fetches without license
When audit logging is enabled and the license has expired or is absent,
then any attempt to fetch data from the audit log will be denied.

Note that Redpanda will continue to produce data to the audit log even
when the license has expired so customers will be able to see their
historical audit data once a valid license is set.