iceberg: fix manifest partition_spec serialization
The Iceberg spec expects only the fields to be serialized in the
'partition-spec', rather than the entire partition spec. This code was
originally modeled after pyiceberg, which has since updated its
implementation to correctly follow the spec.

Along the way, I've updated the test artifact generated by pyiceberg,
using pyiceberg 0.7.1, which includes a fix for the same bug in
pyiceberg.