`storage`: add call to `segment_index::flush()` in `mark_seg...`
A call to `maybe_set_clean_compact_timestamp()` should be followed
by a call to `flush()` in order to persist the clean compact timestamp to disk.

Without a call to `flush()`, a broker failure could potentially make
compaction/tombstone removal less efficient than it could be (as the earliest
clean compact timestamp could be lost).

Make `mark_segment_as_finished_window_compaction()` async and add a call to
`flush()` in case it is required (`maybe_set_clean_compact_timestamp()` now
returns a boolean indicating whether or not the timestamp was set).