datalake/translation: fix data partition to coordinator mapping
An inherent assumption in coordinator implementation is that all
topic/table related operations are performed by the same coordinator for
correctness (so we have one active writer per table active at most
times). This is particularly relevant in filesystem catalog
implementation as it relies on the filesystem for concurrency and
multiple interleaved writers can leave it in a questionable state.

This expectation is violated in our coordinator hash function that
factors in topic and partition as input to the hash function resulting
in ntps from same topic hashing to different coordinator partitions.