bytes: remove `_verify_shard` from iobuf
Since seastar::temporary_buffer has been thread-safe via atomic
reference counting iobuf is now transitively thread-safe as well as long
as it's moved to another core and not passed by reference. Therefore
there is no longer a need to ensure an iobuf is always used on the same
core in which it was created.