rpc: don't wait for the obsolete feature on start
That's feature::rpc_transport_unknown_errc.

On boostrap, when a node doesn't have all RPC services available yet, it
may return non-retriable errors on RPC requests. There is already a
mechanism thatmakes RPC return a retriable error on unknown method if
some services are still yet to start. Unfortunately, this mechanism
requires all nodes to support it, so we check for a feature flag before
we use it. As a result, non-retriable errors still can sneak through, as
checking the feature flag across all nodes takes some time.

This is to remove the feature flag check, as current and previous
versions definitely support it.