parquet: remove redundant field for path vs name.
The name is just the last segment in the path, so instead only store the
path and give a computed view of the name available for usage.

This is a little less clear to me the relationship here between what
should be constructed and what is the final data structured. At one
point I had seperate data structures for "what to construct" vs "indexed
schema", maybe there is enough differences to justify that switch.