c/backend: discard local replica state if cloud state is more up to date
When force recovering partition the controller backend logic decides
which partition replica should be a voter and by that which partition
replica can become a leader. The default policy in controller backend
added all survivor replicas to the voter set, this way promoting their
local state as the source of truth for the partition data.

In some rare scenarios it may be the case that the survivor local state
is 'behind' the state from the cloud. In this case controller backend
should recognize that is is more beneficial to use cloud data as a
source of truth.

This commit adds a controller backend logic that reconciles with the
cloud and selects whichever state is more up to date when executing
partition force reconfiguration. If cloud offset is greater than the
replica dirty offset, the replica local state is removed and cloud
recovery is executed, this way the more up to date state is recovered
and data loss after critical failure is minimized.

Signed-off-by: Michał Maślanka <michal@redpanda.com>