cluster: improve role-acl create sanction message
When a user tries to create a Role-bound ACL, we return an
UNKNOWN_SERVER_ERROR to the Kafka API request and print this log
message.

This log message should clarify that the reason for rejecting the ACL
creation request was that the user was trying to create a Role-bound ACL
without a valid enterprise license.