dt: increase redpanda start timeout to 40s
In Azure CDT tests we are seeing that starting redpanda occasionally
takes longer than 30s. This seems to be because of disk access, in
particular creating directories, being slow while redpanda is starting
up.

We have increased the timeout from 20s to 30s in the past for this
reason, but we are still seeing startup taking longer than 30s, so this
increases the timeout further to 40s.

Fixes https://redpandadata.atlassian.net/browse/CORE-8184