datalake/tests/services: fix credential management in CDT
CDT tests run with IAM roles attached to the instance. In such cases we
don't need to plumb the keys and instead let the AWS sdk pickup the
temporary credentials from the IAM role.

This commit makes the key configs optional which are optionally only
loaded in docker setup with minio while avoiding them in a dedicated
node context.