dl/coordinator: deduplicate files committed to Iceberg
Updates the Iceberg file committer to deduplicate files that have
already been committed. It does so by adding the currently committed
control topic partition offset with each new Iceberg snapshot. If
leadership changes and we attempt to commit repeated files, this offset
tells the coordinator which files are repeats and can be skipped.