auditing: construct metrics probe during service construction
Move the construction of the audit probe from service start to service
construction. Previously, the audit probe was created as a unique pointer
when the service was started and destroyed when the service stopped.

Since the audit service has the same lifespan as the application, creating
the probe during service construction reduces the risk of dereferencing a
null pointer if the probe is accessed before the service has started.

Signed-off-by: Michael Boquard <michael@redpanda.com>
(cherry picked from commit 862e549741b0c196b9a5e0e577cb6cf2137ee677)