c/tm_stm: refactored error handling when expiring transactions
Previously logic in tm_stm verified if the transaction was removed from
the transactions map after it was replicated. This logic caused a lot of
spurious warning logs and was very hard to understand. Additionally the
check is not required as now all the updates of the transactions cache
are replicated. Simplified the logic by skipping the check of the
transaction cache when replicating transaction tombstone. This way a
warning message will only be logged when the regular transaction state
update refers to not existing transaction.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 5f8d126fb2c3ef38ff9e7473abfbdf02a7687e6f)