cluster: introduce simple_allocation_request code path
Create a new partition allocation path. The new path doesn't explicitly
instantiate a full allocation_request upfront, as the size of this request
scales with the number of requested partitions. This allows to check for
system limitations and reject a request earlier, without the need to
create a potentially resource-intense allocation_request.