datalake/translator: added ability to use serde parquet writer
Added an environment variable that will allow us to use serde parquet
writer in tests. The variable is going to be removed when we are going
to be ready to remove the dependency on the arrow library.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit dc03d0c196468542dfba6974bf9de0c84f2eb583)