audit: clamp audit client max parallelism
This fixes a bug in the audit client where if the cluster config value
`kafka_batch_max_bytes` is greater than `audit_client_max_buffer_size`,
the audit client ends up not producing any messages and fills up the
audit log buffers.

The problem is the division here could lead to `max_concurrency=0`. In
debug mode this is caught by an assert inside
`ss::max_concurrent_for_each` but in release mode stdlibrary `assert`s
are disabled and the behaviour ends up being that the background fibre
just blocks waiting for a semaphore unit to be available, but it never
will become available because it was initialized to 0 and no tasks will
ever release any units to it.