parquet: add zstd compression
Add the ability to compress parquet column chunks using zstd.

We currently only support zstd because that is the defacto compression
format that seems to be used these days and it's the only compression codec
that we have futurized as to not cause stalls.

(cherry picked from commit cab8e0f3eba4ed503bef11a3098dae09f5fb79ce)