dt/license: fix expected license response in fips
We recently changes the response format of the cluster license info
endpoint to include the list of enterprise features used. The test was
fixed up for normal non-fips mode execution, but in fips mode the test's
expectation of the response is incorrect since it should expect to see
fips as one of the enterprise features used.