datalake: update default table fields
Updates the default table fields:
- wraps all system fields in a top level `redpanda` nested struct
- unwraps user fields instead of having them all be in a nested struct

This isn't the final form of default schemas, as we currently still
allow going from "schemaless" to schema-ed, while the desired end state
is to use topic configs to force users to choose. Until that is
implemented, the `redpanda` struct still contains the `value` field to
allow for going between the two modes.