rptest/context: add cloud_storage
The Context package contains "accessors" for Ducktape's [TestContext][1]
tailored for our domain and that can be shared by different services or
that can used directly by specific tests.

cloud_storage package provides typed access to cloud storage
configuration. Note that the bucket is not available from the
credentials that is because the bucket is actually created by the
redpanda service. We'll probably refactor this a little bit later but
for now that's enough as bucket is uniform across all cloud
providers/storage type so we can pass it as is.

[1]: https://github.com/confluentinc/ducktape/blob/6136efbb84710f7f7a200172ae88873ab154582d/ducktape/tests/test.py#L262